src/main/test/java/com/aoc/y2024/ProblemDay19Test.java
package com.aoc.y2024;

import com.aoc.ProblemDayTest;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.fest.assertions.Assertions.assertThat;

public class ProblemDay19Test implements ProblemDayTest {

    private ProblemDay19 problem;

    @BeforeEach
    public void before() {
        problem = new ProblemDay19();
    }

    @Test
    @Override
    public void testSolveA() {
        assertThat(problem.solveA()).isEqualTo(278);
    }

    @Test
    @Override
    public void testSolveB() {
        assertThat(problem.solveB()).isEqualTo(569808947758890L);
    }

}
